fix(ci): split strategy attempt hotspots
