+++ native/rust/crates/ripdpi-monitor-engine/src/engine/runners/strategy.rs
@@ -1,3 +1,4 @@
+mod attempt_recording;
 mod connection_concurrency;
 mod dns_baseline;
 mod quic;
