+++ native/rust/crates/ripdpi-monitor-engine/src/engine/runners/strategy/tcp/baseline.rs
@@ -3,18 +3,16 @@ use std::sync::Arc;
 use rustls::client::danger::ServerCertVerifier;
 
 use crate::candidates::StrategyCandidateSpec;
-use crate::classification::{classified_failure_probe_result, classify_strategy_probe_baseline_observations};
-use crate::execution::{CandidateExecution, StrategyLaneExecutor};
-use crate::observations::observations_for_results;
-use crate::types::{DomainTarget, StrategyProbeAttemptRound, StrategyProbeProgressLane};
+use crate::types::{DomainTarget, StrategyProbeProgressLane};
 
 use super::super::super::super::runtime::{ExecutionPlan, ExecutionRuntime, ExecutionStageRunner, RunnerArtifacts};
+use super::super::attempt_recording::record_baseline_candidate_attempts;
 use super::super::support::strategy_probe_live_progress_with_targets;
 use super::StrategyTcpRunner;
 
 pub(super) struct BaselineRun<'a> {
     pub(super) spec: &'a StrategyCandidateSpec,
-    pub(super) execution: CandidateExecution,
+    pub(super) execution: crate::execution::CandidateExecution,
     pub(super) hostfake_family_succeeded: bool,
 }
 
@@ -38,7 +36,8 @@ pub(super) fn run_baseline_candidate<'a>(
         baseline_spec.label,
         format!("Testing TCP candidate {}", baseline_spec.label),
     );
-    let mut baseline_execution = runner.lane_executor.execute_tcp_candidate(
+    let mut baseline_execution = crate::execution::StrategyLaneExecutor::execute_tcp_candidate(
+        &runner.lane_executor,
         baseline_spec,
         domain_targets,
         strategy_plan.runtime_context.as_ref(),
@@ -47,19 +46,17 @@ pub(super) fn run_baseline_candidate<'a>(
         plan.request.diagnostic_tls_keylog_path.as_deref(),
         runtime.cancel_token(),
     );
-    for attempt in &mut baseline_execution.attempts {
-        attempt.round = StrategyProbeAttemptRound::Baseline;
-    }
-    runtime.strategy.attempts.append(&mut baseline_execution.attempts);
+    record_baseline_candidate_attempts(runtime, &mut baseline_execution);
     if baseline_execution.cancelled {
         return None;
     }
 
-    let baseline_observations = observations_for_results(&baseline_execution.results);
-    runtime.strategy.baseline_failure = classify_strategy_probe_baseline_observations(&baseline_observations);
+    let baseline_observations = crate::observations::observations_for_results(&baseline_execution.results);
+    runtime.strategy.baseline_failure =
+        crate::classification::classify_strategy_probe_baseline_observations(&baseline_observations);
     let mut baseline_results = baseline_execution.results.clone();
     if let Some(failure) = &runtime.strategy.baseline_failure {
-        baseline_results.push(classified_failure_probe_result(baseline_spec.label, failure));
+        baseline_results.push(crate::classification::classified_failure_probe_result(baseline_spec.label, failure));
         // Publish the DPI failure class so the UI can show it as a badge.
         runtime.publish_progress(
             plan,
@@ -94,7 +91,6 @@ pub(super) fn run_baseline_candidate<'a>(
         ),
     );
 
-    let hostfake_family_succeeded = baseline_execution.summary.family == "hostfake"
-        && baseline_execution.summary.succeeded_targets == baseline_execution.summary.total_targets;
+    let hostfake_family_succeeded = baseline_execution.fully_succeeded_family("hostfake");
     Some(BaselineRun { spec: baseline_spec, execution: baseline_execution, hostfake_family_succeeded })
 }
