+++ native/rust/crates/ripdpi-monitor-engine/src/execution/lanes/https/sample_builder.rs
@@ -8,8 +8,7 @@ use crate::candidates::StrategyCandidateSpec;
 use crate::execution::scoring::ProbeSample;
 use crate::tls::TlsKeyLogCallback;
 use crate::transport::TransportConfig;
-use crate::types::{DomainTarget, ProbeDetail};
-use crate::util::now_ms;
+use crate::types::DomainTarget;
 
 use super::detail_builder::build_https_probe_details;
 use super::observation_collection::collect_https_observations;
@@ -36,21 +35,12 @@ pub(super) fn build_https_probe_sample(
         &outcome,
         &mut details,
     );
-    details.push(ProbeDetail { key: "probeRetryCount".to_string(), value: retry.retry_count.to_string() });
-    let attempt_reason = details
-        .iter()
-        .find(|detail| detail.key == "tlsError" && detail.value != "none")
-        .map(|detail| detail.value.clone());
-    let duration_ms = now_ms().saturating_sub(observations.started_at_ms);
-
     sample_result::build_https_sample(
         candidate,
         target,
         retry.final_outcome,
         details,
         observations.started_at_ms,
-        duration_ms,
         retry.retry_count,
-        attempt_reason,
     )
 }
