+++ native/rust/crates/ripdpi-monitor-engine/src/execution/scoring/tests.rs
@@ -67,11 +67,7 @@ fn candidate_score_add_accumulates_weighted_success() {
         success: true,
         weight: 2,
         quality: 4,
-        latency_ms: 50,
-        started_at_ms: 100,
-        retry_count: 0,
-        protocol: "TEST".to_string(),
-        reason: None,
+        attempt: ProbeAttemptMetadata::new(100, 50, 0, "TEST", None),
         domain: None,
         is_control: false,
     });
@@ -85,11 +81,7 @@ fn candidate_score_add_accumulates_weighted_success() {
         success: false,
         weight: 1,
         quality: 0,
-        latency_ms: 100,
-        started_at_ms: 200,
-        retry_count: 0,
-        protocol: "TEST".to_string(),
-        reason: Some("failed".to_string()),
+        attempt: ProbeAttemptMetadata::new(200, 100, 0, "TEST", Some("failed".to_string())),
         domain: None,
         is_control: false,
     });
@@ -116,11 +108,7 @@ fn candidate_score_full_success_when_all_targets_succeed() {
         success: true,
         weight: 1,
         quality: 3,
-        latency_ms: 100,
-        started_at_ms: 100,
-        retry_count: 0,
-        protocol: "TEST".to_string(),
-        reason: None,
+        attempt: ProbeAttemptMetadata::new(100, 100, 0, "TEST", None),
         domain: None,
         is_control: false,
     });
@@ -141,11 +129,7 @@ fn candidate_score_preserves_control_classification_in_domain_outcome() {
         success: true,
         weight: 1,
         quality: 3,
-        latency_ms: 25,
-        started_at_ms: 100,
-        retry_count: 0,
-        protocol: "TEST".to_string(),
-        reason: None,
+        attempt: ProbeAttemptMetadata::new(100, 25, 0, "TEST", None),
         domain: Some("control.example".to_string()),
         is_control: true,
     });
@@ -269,11 +253,7 @@ fn build_execution_preserves_target_attempt_timing_and_retry_evidence() {
         success: false,
         weight: 2,
         quality: 0,
-        latency_ms: 250,
-        started_at_ms: 1_000,
-        retry_count: 1,
-        protocol: "HTTPS".to_string(),
-        reason: Some("operation timed out".to_string()),
+        attempt: ProbeAttemptMetadata::new(1_000, 250, 1, "HTTPS", Some("operation timed out".to_string())),
         domain: Some("blocked.example".to_string()),
         is_control: true,
     });
